pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ngf-backend"
version = "0.1.0"
description = "Next Generation Fitness backend scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "pydantic-settings>=2.8.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "sqlalchemy>=2.0.38",
  "sqlalchemy-utils>=0.41.2",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
